rag-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
qdrant-client==1.10.1
sentence-transformers==3.0.1
whoosh==2.7.4
watchfiles==0.22.0
pypdf==4.3.1
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.9.1
tiktoken==0.7.0
transformers==4.44.2
torch==2.4.1
einops==0.8.0
safetensors==0.4.5
PyJWT[crypto]==2.9.0
